Split out the TOTPGenerator utility into its own file
